+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.20.0
 
 * `[system/process]` Added methods for setting and getting CPU scheduler info
+* `[system/process]` Added methods for setting and getting IO scheduler info
 
 ### 12.19.0